+++ app/core/login/templates/loginrecpass.html
@@ -49,7 +49,7 @@
           </form>
 
           <p class="mt-3 mb-1">
-            <a href="/login" class="link-estilo">Login</a>
+            <a href="/core/login" class="link-estilo">Login</a>
           </p>
           <p class="mb-0">
             <a href="/registro" class="text-center link-estilo">Registrar nuevo usuario</a>
