+++ app/core/login/templates/loginregusu.html
@@ -89,7 +89,7 @@
     
           
     
-          <a href="/login" class="text-center link-estilo">Ya estoy registrado</a>
+          <a href="/core/login" class="text-center link-estilo">Ya estoy registrado</a>
         </div>
         <!-- /.form-box -->
       </div><!-- /.card -->
