+++ app/core/models.py
@@ -1,24 +1,6 @@
 from django.db import models
 
 
-# Modelo/tabla para llevar conteo de registro, podría permitir que un usuario recupere su cuenta
-# Tabla de resgistro de usuario
-class RegUsu(models.Model):
-    nomComUsu = models.CharField(max_length=100, unique=True, verbose_name='Nombre')
-    e_mailUsu = models.EmailField(unique=True, verbose_name='E-mail')
-    passwordUsu = models.CharField(max_length=10, verbose_name='Contraseña')
-    conf_passwordUsu = models.CharField(max_length=10, verbose_name='Confirmar Contraseña')
-    accept_terms = models.BooleanField(default=False)
-
-    def __str__(self):
-        return self.e_mailUsu
-
-    class Meta:
-        verbose_name = 'Registrar usuario'
-        verbose_name_plural = 'Registrar Usuarios'
-        ordering = ['id']
-
-
 # ---No se crea Modelo/tabla para recuperar ocntraseña o iniciar sesión por obvias razones.---
 
 # Modelo/tabla de modulo contactar desarroladores
