+++ app/templates/ayuda.html
@@ -96,7 +96,7 @@
                   </li>
                   <li>
                   <p>¿Como puedo recuperar mi contraseña?</p>
-                  <p>Para recuperar tu contraseña, Ingresa al siguiente enlace <a href="loginrecpass.html">Haga clic aquí</a>
+                  <p>Para recuperar tu contraseña, Ingresa al siguiente enlace <a href="../core/login/templates/loginrecpass.html">Haga clic aquí</a>
                   Ingresas un correo electrónico a donde te llegara un mensaje de recuperación.
                   Esperamos que te hayan servido de mucha ayuda.</p>
                   </li>
