+++ app/templates/headerSimpleNeedlebox.html
@@ -34,7 +34,7 @@
           </li>
 
           <li class="navbar-item">
-            <a href="/login" class="nav-link">CERRAR SESIÓN</a>
+            <a href="/core/login" class="nav-link">CERRAR SESIÓN</a>
           </li>
 
         </ul>
